feat: #426 extract registration modal into separate file
