+++ resources/js/Components/Modal/index.ts
@@ -1,0 +1,1 @@
+export {default as EventRegistrationModal} from './EventRegistrationModal'
